+++ radiomesh/literals.py
@@ -4,11 +4,7 @@ from typing import Any, Callable, Generic, Hashable, Tuple, TypeVar
 
 from numba.core import types
 from numba.core.datamodel.models import OpaqueModel, register_default
-from numba.extending import (
-  NativeValue,
-  typeof_impl,
-  unbox,
-)
+from numba.extending import NativeValue, overload_attribute, typeof_impl, unbox
 
 
 class Schema(tuple):
@@ -115,6 +111,11 @@ class DatumLiteral(Generic[H], types.Literal, types.Dummy):
     return self.literal_value.value
 
 
+def is_datum_literal(obj, typ):
+  """Return True if obj is a DatumLiteral holding a Datum of the given typ"""
+  return isinstance(obj, DatumLiteral) and isinstance(obj.datum_value, typ)
+
+
 @unbox(DatumLiteral)
 def unbox_datum_literal(typ, obj, c):
   """Convert a Python DatumLiteral to a Numba representation
@@ -136,3 +137,10 @@ register_default(DatumLiteral)(OpaqueModel)
 
 # This ensures numba.literally(Datum(...)) produces a DatumLiteral
 types.Literal.ctor_map[Datum] = DatumLiteral
+
+
+@overload_attribute(DatumLiteral, "value")
+def overload_datum_value(self):
+  """Returns the literal_value of a DatumLiteral"""
+  value = self.datum_value
+  return lambda self: value