+++ radiomesh/tests/test_literals.py
@@ -6,7 +6,11 @@ from llvmlite import ir
 from numba.core.errors import RequireLiteralValue
 from numba.extending import intrinsic, overload
 
-from radiomesh.literals import Datum, DatumLiteral
+from radiomesh.literals import Datum, DatumLiteral, is_datum_literal
+
+
+def test_is_datum_literal():
+  assert is_datum_literal(DatumLiteral(Datum(4.0)), float)
 
 
 def test_datum_literal_name():
@@ -63,3 +67,14 @@ def test_datum_literal():
   passed through njit, overloads and intrinsics"""
 
   assert f(1.0, Datum((2, 3, 4))) == 10.0
+
+
+def test_datum_literal_jit():
+  value = 4.0
+  datum = Datum(value)
+
+  @numba.njit
+  def fn():
+    return datum.value
+
+  assert fn() == value